Moved bodyparser & errorhandler to a common package
