+++ internal/fiberparser/fibererrorhandler.go
@@ -1,10 +1,10 @@
-package fibererrorhandler
+package fiberparser
 
 import (
 	"github.com/gofiber/fiber/v2"
 )
 
-// Register registers a new Fiber Error Handler. It needs to be used with the Panic MW
+// RegisterErrorHandler registers a new Fiber Error Handler. It needs to be used with the Panic MW
 //
 // It overrides the REST error response when the application has an unexpected error
 // (to avoid information disclosure to the client).
@@ -13,7 +13,7 @@ import (
 //
 // (More info on: https://docs.gofiber.io/api/middleware/recover and https://docs.gofiber.io/guide/error-handling)
 //
-func Register(ctx *fiber.Ctx) error {
+func RegisterErrorHandler(ctx *fiber.Ctx) error {
 	// Set Content-Type: text/plain; charset=utf-8
 	ctx.Set(fiber.HeaderContentType, fiber.MIMETextPlainCharsetUTF8)
 	return ctx.Redirect("/")
